FAI_Pro/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
langgraph==0.1.19
langchain-core>=0.2.22,<0.3
groq==0.9.0
pandas==2.2.2
numpy==1.26.4
python-dotenv==1.0.1
pydantic==2.7.4
httpx==0.27.0
tavily-python==0.3.3
